Use stream v0 when probing for color space
When a file contains both a video and a picture, for example when saving thumbnail with ytarchive, the script finds the color space from the thumbnail and applies it to the video. If the picture contains a different color space it will make the final video have bad colors.

Fix will select video stream so it can find the correct color space.